refact: update folder structure
