+++ src-tauri/src/commands/mod.rs
@@ -1,0 +1,3 @@
+pub mod resources;
+pub mod server;
+pub mod window;
