+++ src-tauri/src/error.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+use thiserror::Error as ThisError;
+
+#[derive(ThisError, Debug, Serialize, Deserialize)]
+pub enum Error {
+    #[error("The local ip is not avaliable.")]
+    LocalIpAddressNotFound,
+}
