+++ src-tauri/src/server/endpoints/sounds.rs
@@ -1,4 +1,4 @@
-use kenku_control::{playlist::playback::*, Controller};
+use kenku_control::Controller;
 use rocket::{get, put, State};
 use serde_json::{json, Value};
